requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-vision==0.40.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
picamera==1.13
pkg-resources==0.0.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
